assignment/requirements.txt
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
decorator==5.1.1
Django==4.0.2
djangorestframework==3.13.1
executing==0.8.2
idna==3.3
ipdb==0.13.9
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
